Adding specific width/height
Adding a specific width and height to both the size of the sketch, as well as the constructor of the ketaiCamera. Up until now we had been trying to use the global width and height variables to set the size of the camera, but this seems to have been what was blocking it from running on the College Nexus tablets. Currently, the sketch is now running on the tablet, with one strange different - on Eiren's phone, the default camera seems to be the back camera, whereas on the nexus, the default seems to be the front facing one.