Responsive to Device Size
Making the app responsive to the device size by setting the size() to (displayWidth, displayHeight). Unfortunatly, the constructor for the Ketai camera does not want to accept the global width/height or dispalyWidth/displayHeight properties as arguemtns, so instead I have created the appWidth/appHeight variables, to store the value of displayWidth/displayHeight and then use these to pass to the ketaiCamera constructor. Ideally, it would be nice if we could avoid this additional step (of passing one variables value to another, just to use it somewhere else) but at the moment it appears that this is the only way to get the camera to display as fullscreen on the college Nexus tablets. At the moment, the app is locked to landscape orientation, as when it is left to auto the image is not responding to the new orientation of the device.