Re-exporting APK with new Version Number
Re-exporting the most recent APK with a new version number, as when I tried to previous one (with the fix for Google Nexus Tablets) to Google Play, I got the error 'You need to use a new version code for your APK because you already have one with version code 1'. To change this I updated the value of the 'android:versionCode' property in the sketch's 'AndroidManifest.xml' file to '2', and changed the 'android:versionName' to '1.1' (as this is not a new release, but a fix of a previous one). The sketch_160118_a_testingCameraKetai-release-signed-aligned.apk has now been uploaded to Google Play and will be available in the next few hours