Changing min and target APK values
Attempting to resolve the issue of Google Play Store not allowing the college Nexus 7 tablets to download. In the 'AndroidManifest.xml' file I am resetting the values of the 'android:minSdkVersion' to 15 and 'android:targetSdkVersion' to 22. Then replacing the line '<uses-permission android:name='android.permission.CAMERA'/>' with '<uses-feature android:name='android.hardware.camera' android:required='false'/>' because there appears to be an issue with the terminology that determines whether the Google Nexus 7's from 2012 are compatible or not. This is just a theory. Version 1.3 (sketch_160118_a_testingCameraKetai-release-signed-aligned.apk) has been uploaded to Google Play store. Waiting for update to be processed on Google Play to test it out.