Accessing front facing camera
Using the Ketai library to print out the details and number of cameras available on this device. I am then using the number of cameras to determine which camera to default to. From what I have read, the back camera tends to be referred to as 0, and the front camera tends to be referred to as 1, but as the college nexus tablets only have one camera, their front one is referred to as 0. This is why I have to check how many cameras there are first.