Attempting to Switch between cameras
Attempting to switch between cameras on an android phone (as the college tablets only have one camera). Currently, this is not working, although I feel that this due to a permissions issue. The app requests the use of 'CAMERA' but it doesn't say anything about cameras (i.e. maybe the front facing camera requires a seperate permission? Will look into