Resetting background to black
Resetting the background to black in between each frame i.e. everytime the draw function is called the background is cleared before the frame is displayed. This makes the switch between cameras feel much more instant (no residue is left over from the previous image) and also makes it clearer what sections of the image are missing - because the portrait image does not yet rotate.