Saving Images to device
Trying to save an image from the device when there is a key event
occurring. This appears to work in the console but currently doesnt.
There is a new folder on the device in the external storage with the
name of the sketch as a folder name on the Phones Pictures folder but no
actual image is being saved.

Calculated photo path:
/storage/emulated/0/Pictures/sketch_160118_a_testingCameraKetai/ourApp.jpg
The console is tracing this out. My theory is as it is emulated is one
reason why it isn't working correctly and the other may because I am not
using a : save dir, but Im not to sure yet.