Storing images -  storage directory
Correction it has created that folder in the internal device storage.
This is a problem as when you store anything on the internal storage of
a device if the app was unistalled by the user all data relating to this
app would also be removed, when we don't want that to be the case and
want to allow the user to hold onto any photos they may have taken.
Currently having trouble with storing an actual image. I am able to
create a directory but the image itself isnt being created.