Moving Icon Loop to Screen Class
Delcaring the allIcons[] array within the Screen class, and moving the loop for showing and checking icons into here aswell. The reason behind this is that now all individual screens can just declare however many icons they would like, and then let the Screen class take care of showing them, and checking if they were clicked on. This meant that i was able to pull functions like 'showHomeScreenPage' out of each instance of the Screen class, and allow them each to inherit the showScreen() method, which takes care of showing the screen, showing the icons, and checking if they were clicked on.