Createing the setScreenIcons Method
Creating a setScreenIcons() method, to pass the temporary allIcons array from each of the screens into the Screen class. Up until now, I had been declaring the allIcons array in the Screen class (as a protected variable) and then allowing each subclass to initialise this array within their own class. The reason I am changing this now is that I feel it is more correct to force the subclasses to use set/get methods where possible, so that later on I can implement testing on the parametres they pass in i.e. to ensure that the array that is passed in has a length greater than 1 etc