Removing Rotation Property from Constructors
Removing the rotation property from all constructors, and using a setRotation() method instead. This is in view to reducing the complexity, and amount of repeating data being passed in to each instance's constructor i.e. if I just default each object to have a rotation of 0, then there is no need for the user to specify it unless they need to later