Creating a Template Screen
Creating a templateScreen that contains all of the comments relevant to the basics of each screen, as well as a homeIcon added to the top right of the screen. Also went through the Icon class's constructors, and removed the ones that were not currently being used, so there are now two custom ones that set defaults for certain parametres if they are not specified, and one full one that sets all of the parametres.