Adding Icon Positioning Variables
Adding icon positioning variables globally to the main sketch i.e. iconLeftX, iconTopY. These will be used when creating icons such as the home icon, or any of the ones on the camera live view, to ensure consistency between screens. Currently, all other icons (such as on the home screen) will need to be positioned using custom values, as their placement is not consistent with other screens.