Trying to streamline displaying of Screens
Trying to reduce the amount of unnecessary code within each screen's class i.e. up until now, each screen had to contain a showScreen(), even if it didn't have any additional parts to be drawn other than the icons, just to call it's super class's drawScreen() method. Currently, I can test if a class contains a screenAdditions() method, but even if it does, I can't currently call it. On the plus side, I now have the screen's switching much more efficently in the main sketch draw function. Each screen is stored in the allScreens array when the app is opened, and after that I just loop through this array to find the class name that matches the currentScreen string (which is set in the Icon class, to reflect the screen which the icon that was clicked should link to.