Removing Screen Instance Names
In the main sketch, instead of creating each screen with an instance name i.e. myHomePage, I am instead passing the screen's constructor directly into an index position in the allScreens array. The reason I have chosen to do this now, is that since the screen's are no longer being called directly in the draw function, but instead being referred to as index positions of the allScreens array, there is no longer a need for them each to have their own instance name as they will only ever be referred to through this array.