Using Switch Statement to Change Screens
Decided to rework the way in which screens were being changed. Instead of storing all of the screens in an allScreens array and looping through it, I am now just calling the right screen using a switch statement. The reason for this change is that in Processing, each array can only hold one datatype. Up until now I was using Screen[] as the datatype for the array, as this is the class from which all screens extend. The problem with this was that even through each screen was being declared as it's own datatype (i.e. HomeScreen) before it was passed into the array, I was not able to access any methods of that screen's own class later on (i.e. when looping through it in the draw function) because it was been seen as a Screen object, and so could only access methods declared within the Screen, or it's super class Rectangle. By going back to a switch statement, even though it looks bulky in the main sketch, I can now allow each screen to be based off the same template (Screen) but still be able to have it's own functionality built in.