Creating an Appropriate directory
Creating an Appropriate directory in the pictures folder of the internal
storage. Awaiting an my update to complete of my SDK so I can access the
android.os.environment so the image can be saved into the external
storage rather than internal, but it is creating the correct folder in a
good location, now trying to get it to actually save the image. I've
tried a few ways including save, saveFrame, savePhoto and a problem I am
having is trying to get the app to call the onsavephotoevent function
but it doesn't seem to be calling it at the moment