Moving testing timeout code into new funciton
In the main sketch, I am now using the testingTimeoutScreen() method to force a screen that doesn't have any interactions to move on to the next screen. This is for testing purposes only - normally these screens will disappear when the thing they are waiting for i.e. return of search results, has occurred