Changing Switch to If/Else
Changing the switch statement in the main sketch (switchScreens) back to an if/else statement, as when I tested it on the Nexus tablet a moment ago, the error 'Cannot switch on a value of type String for source level below 1.7. Only covertible int values or enum variables are permitted.' Which I have since learned is because using switch statements on String objects was a feature introduced in Java 1.7. but Android requires version 1.5 or 1.6.