Trying to Seperate Icons and Buttons
Trying to seperate Icons and Buttons, by creating one common super class (ClickableElement) and allowing each to be treated seperatley. Beginning to realise that this may result in the same issue as the Screens (i.e. if buttons are seperate to icons, then they would need to be stored in Buttons and Icons arrays, as opposed to in the one array).