Enabling Text Input
Only getting to commit this now due to a corrupted branch. This code is to enable text input. Currently there are text field that allow for input. If a mouseEvent occurs such as mousePressed() the keyboard will pop to allow the user to type. bUt my current problem is getting the key that is pressed to put its value into the text input box or or atleast recognise te different keys. It can detect if a key is pressed, just not a specific one. The keyboard appearing is done using the ketai keyboard and the input fields are Control p5 library