Storing String Values
The sketch is now able to store string text characters. The backspace functionality to remove the last text character is now working. The majority of the keyboard characters work with the exception of a few. The keyboard keys that are currently causing problems are: • √ π ÷ x △ ¶ £ ¢ € ¥ ° © ® ™ ℅ ¡ ¿ - We are most concerned about £ € ¥ ¡ and ¿ as these seem to be common to most keyboards.