Prototyping Saving Edited Images - Option 1
Trying to use loadPixels() and updatePixels() on the ketaiCamera object. The application reponds erratically, and the blue we are trying to place on the right hand side of the image, instead flashes over the entire screen, including behind the image somehow. Based on this test, we have decided that although it appears possible to read in the pixels from the ketaiCamera image, it is not possible to manipulate or update them.