Prototyping Saving Edited Images - Option 2
Trying to use a PImage to manipulate the image from the camera, and add in some blue, and then pass the pixel values of this back to the ketaiCamera before it saves. We can’t make ketai camera equal to another images set of pixels. (ie. using the get() method). This method causes the error - type mismatch 'processing.core.PImage' does not match with 'ketai.camera.KetaiCamera'.