Prototyping Saving Edited Images - Option 3
Trying to use the currentImage (PImage) to display on live view screen and then update ketai camera pixel array, by looping through the ketaiCamera array and setting each pixel one by one to be equal to the equilivant currentImage pixel, before saving. The image is saving but with no change in pixels or no changing colour.