Testing out code on other random green screen images
Currently, it appears to be working on greenScreenImage's 1, 2 and 3. I chose greenScreenImage 4 due to it's heavy shadow behind the person, and currently this is not working well. I have added in some code that fixes it, but have since commented it out, as it breaks the other images. I am considering this the limit of how much shadows our code can currently accomodate for.