Twitter Tweeting
Ability to tweet from the wish i was here account on twitter. Also as part of the Twitter access, still trying to get the user to be authorized but having trouble. I know once I have the access to the user tokens and keys, tweeting should be easy enough but the fact that I am trying with Oauth a lot of online resources are outdated and arent a good indication to find out if I am going about it the right way. I have tried with the twitter4j api code example and I still get errors, going to see if I can get atleast some token or key so atleast i can work on for it