Re-releasing Sprint-005-Alpha-10-Release
The Google Play Store threw the following error when we attempted to upload the new APK - 'This configuration cannot be published for the following reason(s): It is forbidden to downgrade devices which previously used M permissions (target SDK 23 and above) to APKs which use old style permissions (target SDK 22 and below). This occurs in the changes from version 9 (target SDK 23) to version 10 (target SDK 21). Version 9 is not served to any device configuration: all devices that might recieve version 9 would recieve version 10.'                                                 Based on this, we realised it was due to an issue with our change in permissions, and we removed '<uses-permission android:name='android.permission.AUTHENTICATE_ACCOUNTS'/>' from our Android manifest, and the issue was resolved. Sprint-005-Alpha-10 is now live on the Google Play Store (in Alpha testing mode)