Tweeting Images and Text
I have made a prototype sketch to send a tweet with text and an image from the main account, I also found about file sizes, when experimenting I had a file too big, so i got an error 413 which from research I learned that file was too large. When I changed image to something smaller it worked. I have also implemented a character check statement, so if the text is more than 126 characters that the tweet wont send until wrong, I am about to test this as I havent yet, hopefully it will work and stop the tweet from sending if i exceed the character limit