Testing various aspects of tweeting and addition of comments
I began by testing what would happen if the user exceed the character limit I set, and it works, it tells the user that the message was too long so they were unsuccessful. I also tested what would happen if the user decided not to put any text in, with this I found out that the tweet is fine and only the extra text of the #WishIWasHere is added, no errors occur when there is no text. Currently it is not taking the images from aa devices storage but the absolute path that is being set to find the image, so I need to find a way to get the latest picture and tweet it. Each image has a unique name and I might need to temporary store the image then delete as I need to give the user the option to save to their device or not, I have started a new sketch for this new development called tweetingFromStorage