Implementing the Google Street View Image API method
Implementing the use of the Google Street View Image API (https://developers.google.com/maps/documentation/streetview/intro#url_parameters) to get Google Street View images and use them as the background of the camera live view. These images are static, and are returned with the appropriate sizing of the current device. Currently, as searching is not yet enabled, all of these locations are stored in the 'Favourites' screen of the app. I have also implemented the random functionality, to ensure there is always a background image in the camera live view screen i.e. each time you press the 'Random' button in the home screen, a random location will be loaded into the background of the camera live view