Successfully merging images
Merging the background image and the keyed green screen image by creating a new PGraphic and adding the images to it. Storing the ketai camera image in the new image at the right orientation by rotating the matrix of the image itself. Thinking that this could have saved us alot of time early on, as we wouldn't have had to keep dealing with the orientation issue, although if we were to do this to every frame, it may have caused a delay. The app is still running quiet slow, so there is a delay between pressing buttons and the next page loading i.e. clicking the Random button on the homepage, it takes time to load in the image from Google images so there appears to be a delay.