Getting to know the methods and options available for the panorama image
Currently I can enable/disable the user's ability to navigate between photospheres, enable/disable the user's ability to pan around the current photosphere, enable/disable the user's ability to zoom in/out of the image and enable/disable the street names being displayed on the ground of the streets. I am now working on trying to change the initial camera angle of the user's view within the photosphere (as this cannot be set via xml the way the GoogleMap was, I am having to create a new camera, and then transition between the old and the new. Currently, this doesn't appear to be working, but is not throwing any errors. Will look into this further.