Checking if resume after twitter Auth resets app
It seems to. Since the callback is calling the main acitiviy, it seems to be resetting everything in the app i.e. the boolean for twitterCalled, which is initialised to false before the setup function is even called, and set to true just before the user is sent to twitter, is back to false when we return from the twitter auth page.