Imitating an interactive Street View Image
Imitating an interactive street view image, by allowing the user to effect the heading and pitch of the static image they are viewing, and in turn make a new request for a new version of the image. Trying to get better search results by trying to get the results pinned with the type 'point_of_interest'. Currently, I don't think this is making much difference, but I do think that if the user adds the country to their search request makes it return more accurate results so we should look at a way to try and encourage them to do this.