Getting location name / refining interactive images
Getting the name of the location, and displaying it on the camera live view screen. Spent some time refining the interaction between the user and the street view image, to make it smoother and less prone to error i.e. that the direction of movement will never flip upside down etc.