Fixes to searching
Checking if a search for a location returns any results. If it doesn't, then you are taken to the SearchUnsuccessfulScreen, which offers you a choice between searching again or going to a random location. Fixed the random location button on this screen, as it was not yet linking to the getRandomLocation function. Also implementing the searching screen i.e. when a user searches for a location, the screen changes to the searching screen, so that even if there is a time delay, they are aware that the app is active.