Incorporating into a prototype of the main sketch for restricting text
I am using the x and y of the text input rectangle to define the x and y starting positions of the text. In order to restrict the text to the input box the rectMode needed to be set to corners as it was set to something else for other componenets in the sketches but this swicthing mode is only called upon when restricitng the text. without changing rectMode the text doesn't appear in the text box and when testing it often was beyond the screens left.