Creating a basis for the app
Creating a new Android Studio project, which has the Processing library imported and used to create a new fragment. From the main Sketch class, I am initialising an instance of the MyClass class (one I created just for testing purposes). Currently, I can pass a string into the constructor of this class, and println() it from the subclass.