Adding Processing library as .jar
Changing how the Processing library is added to the app, by copying the android-core.zip file from the android mode folder of the Processing sketch folder, and adding it to the app's lib folder (renaming it to android-core.jar) as recommended by Processing (https://github.com/processing/processing-android/wiki/Using-Eclipse-or-Android-Studio-for-Development). Currently I can add text to the screen from the Sketch class, but not from the MyClass subclass (as I get java.lang.NullPointerException at processing.opengl.PGL.loadVertexShader - when trying to use the text() method).