Getting sub-class working
According to a Processing forum post I read (https://processing.org/discourse/beta/num_1145945904.html) the reason the Processing methods were not working in the sub classes were because MyClass had not been properly initialized as an applet. To get the MyClass working, I had to pass 'this' into its constructor, and then use this to store in a new PApplet variable within the class. I then had to begin each Processing statement with this variable i.e. fill() becomes p.fill(). While this works, I am not particuarly happy with it, as it would mean having to replace every instance of a Processing class, throughout the entire app, with p.