Not all global variables carry between classes
Based on some additional testing, it seems as though only the value that is initially declared for a variable is the value which carries i.e. in this example, myNum1 is declared and initialised before setup is called, and it's value is accessible from the MyClass show() method. myNum2 however is declared before setup, and then initialised in the setup() function. This value does not carry over, as it seems to be stuck on the default 0.0. This could seriously complicate some of the functionalities we are trying to achieve.