Initial progress on Migrating app to Android Studio
This is the progress I have made so far on migrating our app to Android Studio. So far, incorporating Processing has not been a straight forward process. 1) The Sketch class (which extends PApplet) has full access to all of the Processing methods and properties. Unfortunatly, classes which extend the Sketch class do not inherit these. Instead, I have had to pass 'this' as an argument to the constructor of each Screen class. The parametre recieved by the class is a PApplet, which I then use to initialise a local PApplet variable (which I have called 'p'). All references to the processing library in every other class of the sketch must be called on the 'p' variable i.e. p.image(). As classes are declared and initialised within each of these Screens (i.e. Icons) I must then pass the 'p' variable to their constructors, so that their classes can in turn access the PApplet of the app. 2) The Processing 'color' datatype is not recognised in Java, and so I have had to remove all instances of it from the app. It turns out the 'color' class is really just an integer, so I am instead using an int to store values such as 0xFFFFFF (white). I cannot yet tell if this is working, as the sketch is not yet fully operational. A full explaination of why points 1 and 2 have occured can be found on https://processing.org/tutorials/eclipse/. 3) Numbers which contain decimal values are not considered float's by default. Instead they are considered to be of the 'double' datatype. This would be ok, except that Processing methods expect float arguments, and not double. For the most part, I have tried to keep the additional casting of these values localised to the higher up super class methods, so that we won't have to remember to explicitly cast arguments as floats. 4) I had originally been importing the Processing library wrong (i.e. I was placing the entire processing library into the main java portion of the project. Instead, I am using a copy of the 'android-core.zip' file, which I sourced from the Android Mode folder of Processing 3's sketch folder, renaming it 'android-core.jar' and placing it in the libs folder of the project (as recommended by https://github.com/processing/processing-android/wiki/Using-Eclipse-or-Android-Studio-for-Development). 5) In Processing, we had been loading some images directly into constructors using loadImage(). In Android Studio, Processing threw and error for this, so had to go back and make a variable to store it. 6) In Processing, we had inadvertantly declared two functions of the same name in the main sketch and the favourites sketch (getRandomLocation()). This has never caused an error until now, but I am glad it did as this was completly unintentional. Currently, the sketch is compiling and running in Android Studio. The current issue I am facing is that variable which are declared globally in the Sketch class, but are not initialised until the settings() or setup() methods are called, appear to pass their default value (i.e. for float, the value would be 0.0) to sub classes, whereas if they were initialised globally (which is not always possible, for example in the case of calculating positions based on the size of the device) their values would carry to the other subclasses.