Set/Get seems to cause variable shadowing
As I was running into issues with migrating the main app (i.e. switching of screens using currentScreen wasn't working). I then starting testing this theory in a basic app, and am more confused than before. If I use a set method to change the value of myNum2 in MyClass, the value of it does not appear to change in the draw() function of the main sketch (using either the identifier or the get method to get its value). It does however change in MyClass (using the identifier and get methods). This is confusing, as the get method returns the current value of myNum2 from the main sketch, so how is this method returning two different values when used in the two classes? It's almost like changing the variable just created a new version of it, to shadow the real value.