Adding progress with main app as a prototype
Adding the progress I have made on the last app since my last commit of it, although this version is after gettting quiet messy (due to the issue with variable shadowing I discussed in my most recent commit). The home screen is working now though, due to passing all the relevant sizing, positioning and image variable directly to the constructor of the screen.