Resolving varialbe shadowing issue
As I was writing up my journal on last night's work, I had an idea of how to pass variables and methods between classes using Processing in Android Studio. Up until now, I was passing 'this' into the constructor of all classes in the main Sketch class, and recieving these parametres as PApplet's in the sub classes, storing them in a variable called 'p'. This was the method suggested by Processing (https://processing.org/tutorials/eclipse/), and meant that I could now access Processing methods, such as fill, within the subclasses by calling p.fill(0). The problem was that I could not find a way to effectively access the properties and methods of the main Sketch class. It appeared as if I was able to, but as discussed in my prototyping last night, I was not actually effecting the values of the original variables, but instead making a shadowed version which contained a different value. I did try calling the properties and methods on the 'p' object, but as this was of type PApplet, this was not possible. The idea I had today was that instead of treating the 'this' parametre as a PApplet, to instead treat it as a Sketch datatype, as the instance of the sketch which it refers to would not only contain the Processing library methods and properties, but it's own methods and properties. This worked!! :). I am now successfully changing the value of myNum2 from within the MyClass class, and it is also updating the Sketch class's value for that variable :).