Moving current working version to prototypes
Moving the current working version of the app in Android studio, to the prototyping folder of our project, as I am about to start implementing the fix I came up with earlier to allow global variables to be accesses from sub classes of the Sketch class. This version does not contain the attempted fixes I was trying to do last night, as that version (MigrateApp6) is after turning into a mess due to my tampering with passing variables through constructors.