Successfully running app in Android Studio
After going through every constructor in the app and changing the PApplet parametre to a Sketch parametre, and then going through every statement in the 23 subclasses to preceed every Processing method, and global variables/methods with 'sketch' the app is now successfully working in Android Studio. Came across one or two minor issues once it was working. 1) I had not set any permissions for the app, so it could not access the internet. I have now added all of the relevant permissions to the AndroidManifest.xml file. 2) The app was appearing smaller than the screen size. Once I removed the default padding which AndroidStudio added to the relative layout (activity_main.xml) by adding a new 'none' value to the dimensions file (dimens.xml) it now fills the screen. 3) I had forgotten to preceed 'defaultTextSize' with 'sketch' in the addText() method of the Rectangle class, so all text elements that did not pass a specific text size were not displaying (as their text size could not be accessed). I have since fixed the. 4) The googleImageHeading and googleImagePitch variables were unable to read in values from the favourites random locations, as (float)(googleImageHeading) was not successfully parsing the String value it was getting to a float. I have since replaced this with Float.parseFloat(googleImageHeading) as this is the more correct way to parse String values to Floats (https://docs.oracle.com/javase/7/docs/api/java/lang/Float.html#parseFloat(java.lang.String)). The only continuing issues are that the screen titles do not appear to be taking on the screenTitleY position, and I have not yet been able to test if the camera etc is working correctly, as this phone does not always display this screen (due to low memory). Will test this on Eiren's phone tomorrow.