Testing a new method for passing the instance of the Sketch class to subclasses
Testing out a method, suggested by my lecturer John, to use static methods to access the instance of the Sketch class from subclasses (to combat the variable shadowing issue I was having over the weekend). This method works, and is a much cleaner way of sharing the PApplet and Sketch class (as one, since Sketch inherits PApplet). This would have saved me having to re-write all of the constructors in the app (twice) :)