Prototyping using Fabric.io with Processing
Prototyping a method to use Fabric.io (to login to Twitter) in conjunction with a Processing sketch. Currently, I have created all of the code relating to Twitter in a new TwitterActivity. When the app runs, the user sees the MainActivity (which contains the Processing Sketch fragment). Once they tap on the screen, I use an intent to switch to the TwitterActivity. Technically, the user would then see that activity's screen, but as I wanted to avoid this I am using the preformClick() method on the twitter loginButton, so that a click is triggered on it just as the activity loads, so the user is taken directly to the Twitter authorisation page. Once the TwitterActivity recieves the data from the login, it stores then in public static variables, and then uses another intent to switch back to the MainActivity (i.e. the Processing Sketch). Within the Processing sketch, I am then printing out the values of these variables (username, userid, access token and secret token).