Trying to incorporate Twitter login into main app
Trying to incorporate Twitter login into the main app. Currently the app is crashing on resume due to a Fatal exception 'out of memory'. After some investigation I feel that this is because resuming the MainActivity is resetting the entire sketch, and that since it is trying to recreate all of the variables etc again there is not enough memory available. Will work on pausing the activity as opposed to stopping it completly.