Possible success with Twitter in Main app
By default the user now logs in to twitter first, and is then taken to the main sketch (as this is the only way I can currently guarantee the main sketch does not get rebuilt and thus there is no crash due to lack of memory). Not sure if this is working as I cannot test it on my device, getting Eiren to check it on her phone.