Attempting to use Processing Server Side
Attempting to use Processing serverside, so that we could potentially send images to the server to be processed (i.e. have the green screen removed) to try and take the load off the mobile devices (as the green screening is significantly slowing down the app). Currently, I have made no progress with this, as the processing-js library is throwing errors regarding the 'window' and 'navigator' being undefined, which is because Processing expects to run in a browser. Instead, I have just added in the comments where I plan to put the code once I get the library to work.