Some success with server side image manipulation
Having some success with server side image manipulation, using Node.js along with the 'jimp' and 'onecolor' packages. Currently, I have the green screen image of the girl loaded on the server, I am using jimp to open the image and run through it's pixels, and then using onecolor to cast the color of the pixels to the relevant values e.g. red, green and blue. I am then loading an image into the processing sketch using loadImage() and passing the url of localhost (as I am running the server locally using Visual Studio Code) and then displaying the resulting image in the sketch. The resulting image (as can be seen in finalImage.png) is extremely distorted (and I have not yet performed any keying to the green screen). All I have done is read in the color of the pixel, and then set the colour of the pixel back to it's own color. I am still unsure of how to send the live camera view images to the server, but I am concerned about how much this method will help us as currently it seems to take approximatley 6 seconds to process this one image.