Delete Dirctory and Content
Still even trying to get ths simplier version to work, I have come to the conclusion that there is an error in the main sketch as the deleteDir is an exact copy of the main sketch and I am trying to just change the folder directory name but once I reach the sharescreenA it all falls apart and the app stops responding. This will need to be looked into furhter. I have checked all the other classes and even ran the main sketch but I keep reaching the same error each time when I am keeping the old directory and the new directory name, I have for any other references and there doesnt seem to be any that would cause this conflict. I even thought it might be that I was missing the keys for twitter and the google browser key but nope, even when I put these in, its the same result, I just dont understand why. In the console I am getting errors such as java.io.FileNotFoundException but the thing is, I am not even using nor references the ileOutputStream  which what the error is pointing to. I am using the fileOutputStream in a completely different sketch for the internal saving but not the deleting of a directory and its contents. Other errors are referencing the outputStream and a null object reference in relation to this but again this is in another file for internal storage not the deleteDir or Main sketch and the last error is stating that the application already exists on device, which yes sometimes will throw an error but I have uninstalled it and ran it and Im still getting the result. Thinking that I will revert back to an old basic save file and try it again there after testing that it is saving it, fingers crossed