Releasing Sprint-09-Beta-015
Renamed the package to processing.test.wish_i_was_here so that it would be accepted as a new release on the Google Play Store (otherwise it would be considered a new project. Also used the existing keystore key we generated for our Processing app, again, so that it would be accepted as a new release as opposed to a new app. Renamed the app by changing the activity labels to all equal 'Wish I Was Here' so that the app appears this way when downloaded onto a device. This version is now released as the first beta version of our app on the Google Play Store (versionCode 15, versionName 9.0).