Fixed issue with colour distortion on server
I was using the HSL color space to create the colour object, and using it's brightness to determine how bright or dark a pixel's colour was, when I really should have been using the HSV colour space and using it's value. Basically, I got my colour spaces mixed up and only realised it when I was writing up my journal this afternoon! Currently, I am changing the green background to a yellow-ish colour, just to test how much I can effect the colour based on it's perception of being part of the green screen. NOTE - because the lightness of the colour is a much lower value than the value (brightness) of a colour, that is why the image was becoming distorted, as I was changing the consistency of the colour by inadvertantly treating the lightness as brightness.